moves network config to a separate package
