+++ network/netconf/connection_manager.go
@@ -1,4 +1,4 @@
-package network
+package netconf
 
 import "time"
