+++ network/netconf/errors_test.go
@@ -1,4 +1,4 @@
-package network
+package netconf
 
 import (
 	"fmt"
